Read this as well
Also wanted to talk about the new background.
Aside from looking awesome, it has important characteristics you should
know.  The grass at the bottom of the fence should ALWAYS be the highest
Y coordination that the pet can go.
***This Y coordinate is equal to 2/3 of the background size.***
The constant AREA_MIN_Y holds this value and its important because
screen sizes will always be different. The purposes of constants are
explained in the code comments.