Progress on runaway feature.
Just need to fix the scenario when you kill the app, and when you repoen
it the timer has completed and the pet should run away.
Desired behavior: screen of runaway notification, go create and go home.
Actual behavior: screen of runaway notification, BUT simultaneously a
new counter is started and is ticking as if already in home, which is
bad.

So we just set the counter to 3 days. And after that the app breaks.
lol.