Got feeding mostly working.
Perfection would be making the pet move to the same place every time
relative to the food. Like, every time you feed the pet the pet moves so
the food is at its feet just to the left of the pet, and then the pet
eats the food.
Also should update so poop timer starts when the food is "eaten".