Updated status stuff. NOTE, DELETED ENERGY.
I made the executive decision to remove the energy level stat. It just
overcomplicates things, and it's completely backwards to try and
implement it using a countdown timer. SO we keep it simple; play with
pet, happiness goes up. Feed pet, hunger goes up. Sleeping is on its own
timer.